requirements.txt
alabaster==0.7.12
arrow==0.17.0
astroid==2.4.2
Babel==2.9.0
beautifulsoup4==4.9.3
bibtexparser==1.2.0
certifi==2020.12.5
chardet==4.0.0
docutils==0.16
fake-useragent==0.1.11
free-proxy==1.0.2
future==0.18.2
idna==2.10
imagesize==1.2.0
isort==5.7.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.8
Pygments==2.7.4
pylint==2.6.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-igraph==0.8.3
pytz==2020.5
requests==2.25.1
scholarly==1.0.4
selenium==3.141.0
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.1
Sphinx==3.4.3
sphinx-rtd-theme==0.5.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
stem==1.8.0
texttable==1.6.3
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
